added separate service class for Admin
