+++ quora-api/src/main/java/com/upgrad/quora/api/controller/AdminController.java
@@ -1,7 +1,7 @@
 package com.upgrad.quora.api.controller;
 
 import com.upgrad.quora.api.model.UserDeleteResponse;
-import com.upgrad.quora.service.business.UserBusinessService;
+import com.upgrad.quora.service.business.AdminBusinessService;
 import com.upgrad.quora.service.entity.UserEntity;
 import com.upgrad.quora.service.exception.AuthorizationFailedException;
 import com.upgrad.quora.service.exception.UserNotFoundException;
@@ -11,19 +11,18 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-
 @RestController
 @RequestMapping("/admin")
 public class AdminController {
 
     @Autowired
-    private UserBusinessService userBusinessService;
+    private AdminBusinessService adminBusinessService;
 
     @RequestMapping(method = RequestMethod.DELETE,
             path="/user/{userId}",
             produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
     public ResponseEntity<UserDeleteResponse> deleteUser(@PathVariable("userId") final String userId, @RequestHeader("authorization") final String authorization) throws AuthorizationFailedException, UserNotFoundException {
-        UserEntity deleteUserEntity = userBusinessService.deleteUser(userId, authorization);
+        UserEntity deleteUserEntity = adminBusinessService.deleteUser(userId, authorization);
         UserDeleteResponse userDeleteResponse = new UserDeleteResponse().id(deleteUserEntity.getUuid()).status("USER SUCCESSFULLY DELETED");
 
         return new ResponseEntity<UserDeleteResponse>(userDeleteResponse, HttpStatus.OK);
