+++ quora-service/src/main/java/com/upgrad/quora/service/business/UserBusinessService.java
@@ -21,24 +21,6 @@ public class UserBusinessService {
         return userDao.createUser(userEntity);
     }
 
-    @Transactional(propagation = Propagation.REQUIRED)
-    public UserEntity deleteUser(final String uuid, final String authorizationToken) throws AuthorizationFailedException, UserNotFoundException {
-
-        UserEntity userEntity = getUserFromToken(authorizationToken);
-        if(!userEntity.getRole().equals("admin")) {
-            throw new AuthorizationFailedException("ATHR-003", "Unauthorized Access, Entered user is not an admin");
-        }
-
-        UserEntity deleteUserEntity = userDao.getUser(uuid);
-        if(deleteUserEntity == null){
-            throw new UserNotFoundException("USR-001", "User with entered uuid to be deleted does not exist");
-        } else {
-            userDao.deleteUser(deleteUserEntity);
-        }
-
-        return deleteUserEntity;
-    }
-
     public UserEntity getUserFromToken(String authorizationToken) throws AuthorizationFailedException {
         UserAuthEntity userAuthTokenEntity = userDao.getUserAuthToken(authorizationToken);
